vestigo_internet/requirements.txt
requests
sqlalchemy
speedtest-cli
pythonping